+++ ios/batch_flutter/Sources/batch_flutter/BatchFlutterPlugin.swift
@@ -4,7 +4,7 @@ import Batch
 
 fileprivate struct Consts {
 	static let BridgeVersionEnvironmentVar = "BATCH_BRIDGE_VERSION"
-	static let BridgeVersion = "Bridge/3.0"
+	static let BridgeVersion = "Bridge/3.1"
 	
 	static let PluginVersionEnvironmentVar = "BATCH_PLUGIN_VERSION"
 	static let PluginVersion = "Flutter/3.0.0"